tests/wpt/tests/css/css-inline/text-box-trim/text-box-edge-valid.html

<!DOCTYPE html>
<title>Tests parsing of the text-box-edge property</title>
<link rel="author" title="Google LLC" href="https://www.google.com/">
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#text-edges">
<meta name="assert" content="text-box-edge supports the full grammar leading | [ text | cap | ex | ideographic | ideographic-ink ] [ text | alphabetic | ideographic | ideographic-ink ]">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value('text-box-edge', 'leading');
test_valid_value('text-box-edge', 'text');
test_valid_value('text-box-edge', 'ex');
test_valid_value('text-box-edge', 'cap');
test_valid_value('text-box-edge', 'text alphabetic');
test_valid_value('text-box-edge', 'text text', 'text');
test_valid_value('text-box-edge', 'cap alphabetic');
test_valid_value('text-box-edge', 'cap text', 'cap');
test_valid_value('text-box-edge', 'ex alphabetic');
test_valid_value('text-box-edge', 'ex text', 'ex');
</script>
